refactor: split mainscreen view models into three files: directed, undirected and abstract
